ddr_control/scripts/requirements.txt
absl-py==0.12.0
actionlib==1.13.2
angles==1.9.13
argcomplete==1.12.3
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
base_local_planner==1.17.1
bondpy==1.8.6
cachetools==4.2.1
camera-calibration==1.15.3
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.4.24
Cerberus==1.3.4
certifi==2021.10.8
charset-normalizer==2.0.7
cloudpickle==1.6.0
colorama==0.4.4
controller-manager==0.19.5
controller-manager-msgs==0.19.5
coverage==5.3
cv-bridge==1.15.0
diagnostic-analysis==1.10.4
diagnostic-common-diagnostics==1.10.4
diagnostic-updater==1.10.4
distro==1.6.0
docopt==0.6.2
docutils==0.18
dronekit==2.9.2
dronekit-sitl==3.3.0
dynamic-reconfigure==1.7.1
flake8==3.9.2
flatbuffers==1.12
gast==0.3.3
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.41.1
gym==0.10.5
h5py==2.10.0
idna==3.3
image-geometry==1.15.0
iniconfig==1.1.1
intelhex==2.3.0
interactive-markers==1.12.0
ipykernel==5.5.5
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joint_state_publisher==1.15.0
joint_state_publisher_gui==1.15.0
jsonpickle==1.5.2
jupyter-client==6.2.0
jupyter-core==4.7.1
kconfiglib==14.1.0
Keras-Preprocessing==1.1.2
laser_geometry==1.6.7
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
MAVProxy==1.8.39
mavros==1.9.0
mccabe==0.6.1
message-filters==1.15.13
munch==2.5.0
nest-asyncio==1.5.1
numpy==1.18.5
nunavut==1.2.1
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkgconfig==1.5.4
pluggy==0.13.1
prompt-toolkit==3.0.19
protobuf==3.15.8
ptyprocess==0.7.0
py==1.11.0
py-cpuinfo==8.0.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydsdl==1.12.0
pyflakes==2.3.1
pygame==2.0.1
pyglet==1.5.15
pymavlink==2.4.15
pyparsing==3.0.6
pyquaternion==0.9.9
pyros-genmsg==0.5.8
pyserial==3.5
pytest==6.1.2
python-dateutil==2.8.2
python-qt-binding==0.4.4
pyulog==0.9.0
PyYAML==6.0
pyzmq==22.1.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.26.0
requests-oauthlib==1.3.0
resource_retriever==1.12.6
rosbag==1.15.13
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.13
roslaunch==1.15.13
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.13
rosmake==1.15.8
rosmaster==1.15.13
rosmsg==1.15.13
rosnode==1.15.13
rosparam==1.15.13
rospkg==1.3.0
rospy==1.15.13
rosserial_python==0.9.2
rosservice==1.15.13
rostest==1.15.13
rostopic==1.15.13
rosunit==1.15.8
roswtf==1.15.13
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.2
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.2
rqt_top==0.4.10
rqt_topic==0.4.12
rqt_web==0.4.10
rsa==4.7.2
rtabmap-ros==0.20.14
rviz==1.14.10
sacred==0.8.1
scipy==1.6.2
sensor-msgs==1.13.1
six==1.15.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
smmap==5.0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tf==1.13.2
tf-conversions==1.13.2
tf-slim==1.1.0
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
toml==0.10.2
topic-tools==1.15.13
tornado==6.1
traitlets==5.0.5
turtlebot3_example==1.2.5
turtlebot3_teleop==1.2.5
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xacro==1.14.10
